spirv_std::arch

Function all_memory_barrier_with_group_sync

pub unsafe fn all_memory_barrier_with_group_sync()

Blocks execution of all threads in a group until all memory accesses have been completed and all threads in the group have reached this call.

This is an exact implementation of AllMemoryBarrierWithGroupSync().

From https://docs.microsoft.com/en-us/windows/win32/direct3dhlsl/allmemorybarrierwithgroupsync