Function spirv_std::arch::atomic_f_add

pub unsafe fn atomic_f_add<F: Float, const SCOPE: u32, const SEMANTICS: u32>(
    ptr: &mut F,
    value: F
) -> F

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

load through ptr to get an original value,

get a new value by integer addition of original value and value, and

store the new value back through ptr.

The result is the Original Value.