spirv_std::arch

Function atomic_load

pub unsafe fn atomic_load<N: Number, const SCOPE: u32, const SEMANTICS: u32>(
    ptr: &N,
) -> N

Atomically load through ptr using the given SEMANTICS. All subparts of the value that is loaded are read atomically with respect to all other atomic accesses to it within SCOPE.