Function spirv_std::arch::ddy_coarse_vector

pub fn ddy_coarse_vector<F: Float, VECTOR: Vector<F, LENGTH>, const LENGTH: usize>(
    component: VECTOR
) -> VECTOR

Returns the partial derivative of component with respect to the window’s Y coordinate. Uses local differencing based on the value of component for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of component for the current fragment. That is, over a given area, the implementation can compute Y derivatives in fewer unique locations than would be allowed by ddy_fine.