spirv_std::arch

Function subgroup_and

pub fn subgroup_and<I: VectorOrScalar<Scalar = impl Integer>>(value: I) -> I

A bitwise and group operation of all value operands contributed by active invocations in the group.

Result Type must be a scalar or vector of integer type.

Execution is a Scope that identifies the group of invocations affected by this command. It must be Subgroup.

The identity I for Operation is ~0.

The type of value must be the same as Result Type.

Requires Capability GroupNonUniformArithmetic.