spirv_std::arch

Function subgroup_ballot_find_msb

pub fn subgroup_ballot_find_msb(value: SubgroupMask) -> u32

Find the most significant bit set to 1 in value, considering only the bits in value required to represent all bits of the group’s invocations. If none of the considered bits is set to 1, the resulting value is undefined.

Result Type must be a scalar of integer type, whose Signedness operand is 0.

Execution is a Scope that identifies the group of invocations affected by this command. It must be Subgroup.

value must be a vector of four components of integer type scalar, whose Width operand is 32 and whose Signedness operand is 0.

value is a set of bitfields where the first invocation is represented in the lowest bit of the first vector component and the last (up to the size of the group) is the higher bit number of the last bitmask needed to represent all bits of the group invocations.

Requires Capability GroupNonUniformBallot.