Function spirv_std::arch::subgroup_logical_and

pub unsafe fn subgroup_logical_and<I: VectorOrScalar<Scalar = bool>>(
    value: I
) -> I

A logical and group operation of all Value operands contributed by active invocations in the group.

Result Type must be a scalar or vector of Boolean type.

Execution is a Scope that identifies the group of invocations affected by this command. It must be Subgroup.

The identity I for Operation is ~0.

The type of Value must be the same as Result Type.

Requires Capability GroupNonUniformArithmetic.