spirv_std::arch

Function vector_extract_dynamic

pub unsafe fn vector_extract_dynamic<T: Scalar, const N: usize>(
    vector: impl Vector<T, N>,
    index: usize,
) -> T

Extract a single, dynamically selected, component of a vector.

Safety

Behavior is undefined if index’s value is greater than or equal to the number of components in vector.