Struct spirv_std::bindless::RenderResourceHandle

#[repr(transparent)]
pub struct RenderResourceHandle(_);

A handle that points to a rendering related resource (TLAS, Sampler, Buffer, Texture etc) this handle can be uploaded directly to the GPU to refer to our resources in a bindless fashion and can be plainly stored in buffers directly - even without the help of a DescriptorSet the handle isn’t guaranteed to live as long as the resource it’s associated with so it’s up to the user to ensure that their data lives long enough. The handle is versioned to prevent use-after-free bugs however.

This handle is expected to be used engine-side to refer to descriptors within a descriptor set. To be able to use the bindless system in rust-gpu, an engine is expected to have created four DescriptorSets, each containing a large table of max 1 << 23 elements for each type. And to sub-allocate descriptors from those tables. It must use RenderResourceHandle to refer to slots within this table, and it’s then expected that these RenderResourceHandle’s are freely copied to the GPU to refer to resources there.

Buffer Type	Set

Buffers	0

Textures	1

Storage textures	2

Tlas	3

Implementations

impl RenderResourceHandle

pub fn new(version: u8, tag: RenderResourceTag, index: u32) -> Self

pub fn invalid() -> Self

pub fn is_valid(self) -> bool

pub fn version(self) -> u32

pub fn tag(self) -> RenderResourceTag

pub unsafe fn index(self) -> u32

Safety

This method can only safely refer to a resource if that resource is guaranteed to exist by the caller. RenderResourceHandle can’t track lifetimes or keep ref-counts between GPU and CPU and thus requires extra caution from the user.

pub fn bump_version_and_update_tag(self, tag: RenderResourceTag) -> Self

This function is primarily intended for use in a slot allocator, where the slot needs to get re-used and it’s data updated. This bumps the version of the RenderResourceHandle and updates the tag.