spirv_std

Module byte_addressable_buffer

Container for an untyped blob of data.

Structs

ByteAddressableBuffer

ByteAddressableBuffer is a view to an untyped blob of data, allowing loads and stores of arbitrary basic data types at arbitrary indices.