Trait spirv_std::derivative::Derivative

pub trait Derivative {
    pub fn ddx(self) -> Self;

    pub fn ddx_fine(self) -> Self;

    pub fn ddx_coarse(self) -> Self;

    pub fn ddy(self) -> Self;

    pub fn ddy_fine(self) -> Self;

    pub fn ddy_coarse(self) -> Self;

    pub fn fwidth(self) -> Self;

    pub fn fwidth_fine(self) -> Self;

    pub fn fwidth_coarse(self) -> Self;
}

Represents a type that can represent the derivation of an operation.

Required methods

pub fn ddx(self) -> Self

Returns the partial derivative of Self with respect to the window's X coordinate. Returns the same result as either Self::ddx_fine or Self::ddx_coarse, selection of which one is dependent on external factors.

pub fn ddx_fine(self) -> Self

Returns the partial derivative of Self with respect to the window's X coordinate. Uses local differencing based on the value of Self for the current fragment and its immediate neighbor(s).

pub fn ddx_coarse(self) -> Self

Returns the partial derivative of Self with respect to the window's X coordinate. Uses local differencing based on the value of Self for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of Self for the current fragment. That is, over a given area, the implementation can compute X derivatives in fewer unique locations than would be allowed by Self::ddx_fine.

pub fn ddy(self) -> Self

Returns the partial derivative of Self with respect to the window's Y coordinate. Returns the same result as either Self::ddy_fine or Self::ddy_coarse, selection of which one is dependent on external factors.

pub fn ddy_fine(self) -> Self

Returns the partial derivative of Self with respect to the window's Y coordinate. Uses local differencing based on the value of Self for the current fragment and its immediate neighbor(s).

pub fn ddy_coarse(self) -> Self

Returns the partial derivative of Self with respect to the window's Y coordinate. Uses local differencing based on the value of Self for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of Self for the current fragment. That is, over a given area, the implementation can compute Y derivatives in fewer unique locations than would be allowed by Derivative::ddy_fine.

pub fn fwidth(self) -> Self

Returns the sum of the absolute values of Self::ddx and Self::ddy as a single operation.

pub fn fwidth_fine(self) -> Self

Returns the sum of the absolute values of Self::ddx_fine and Self::ddy_fine as a single operation.

pub fn fwidth_coarse(self) -> Self

Returns the sum of the absolute values of Self::ddx_coarse and Self::ddy_coarse as a single operation.

Loading content...