Module spirv_std::image::sample_with

Contains extra image operands

Structs

NoneTy

Helper struct that denotes that the type doesn’t exist, analog to Option::None

SampleParams

Helper struct that allows building image operands. Start with a global function that returns this struct, and then chain additional calls. No care is taken to avoid stating multiple operands that, together, make no sense, such as Lod and Grad. Example: image.sample_with(coords, sample_with::bias(3.0).sample_index(1))

SomeTy

Helper struct that denotes that the type does exist and is of type T, analog to Option::Some(T)

Traits

OptionTy

Helper trait to mimic Option<T>, but where the variant are types

Functions

bias

Sets the ‘Bias’ image operand

grad

Sets the ‘Grad’ image operand

lod

Sets the ‘Lod’ image operand

sample_index

Sets the ‘Sample’ image operand