Struct spirv_std::image::sample_with::SampleParams

pub struct SampleParams<B: OptionTy, L: OptionTy, G: OptionTy, S: OptionTy> {
    pub bias: B,
    pub lod: L,
    pub grad: G,
    pub sample_index: S,
}

Helper struct that allows building image operands. Start with a global function that returns this struct, and then chain additional calls. No care is taken to avoid stating multiple operands that, together, make no sense, such as Lod and Grad. Example: image.sample_with(coords, sample_with::bias(3.0).sample_index(1))

Fields
bias: B

‘Bias’ image operand

lod: L

‘Lod’ image operand

grad: G

‘Grad’ image operand

sample_index: S

‘Sample’ image operandy

Implementations

impl<L: OptionTy, G: OptionTy, S: OptionTy> SampleParams<NoneTy, L, G, S>

pub fn bias<B>(self, bias: B) -> SampleParams<SomeTy<B>, L, G, S>

Sets the ‘Bias’ image operand

impl<B: OptionTy, G: OptionTy, S: OptionTy> SampleParams<B, NoneTy, G, S>

pub fn lod<L>(self, lod: L) -> SampleParams<B, SomeTy<L>, G, S>

Sets the ‘Lod’ image operand

impl<B: OptionTy, L: OptionTy, S: OptionTy> SampleParams<B, L, NoneTy, S>

pub fn grad<T>( self, grad_x: T, grad_y: T ) -> SampleParams<B, L, SomeTy<(T, T)>, S>

Sets the ‘Lod’ image operand

impl<B: OptionTy, L: OptionTy, G: OptionTy> SampleParams<B, L, G, NoneTy>

pub fn sample_index<S>( self, sample_index: S ) -> SampleParams<B, L, G, SomeTy<S>>

Sets the ‘Sample’ image operand