spirv_std::image

Trait ImageSizeQuery

pub trait ImageSizeQuery<T, const DIM: u32, const ARRAYED: u32> { }

Marker trait for query size results based on image dimension and arraying.

This trait represents the SPIR-V size query results:

1D images return a scalar

2D/Cube/Rect images return 2 components (Cube returns face width/height)

3D images return 3 components

Arrayed images add one component for the array size