Type Alias spirv_std::image::Image3dU

pub type Image3dU = Image<u32, { _ }, { _ }, { _ }, { _ }, { _ }, { _ }, { _ }>;

A 3d image used with a sampler, containing unsigned integer data.

Aliased Type

struct Image3dU { /* private fields */ }

Implementations

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const ARRAYED: u32, const MULTISAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { _ }, FORMAT, COMPONENTS>

pub fn fetch<I>( &self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED> ) -> SampledType::SampleResultwhere I: Integer,

Fetch a single texel with a sampler set at compile time

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32, const SAMPLED: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, ARRAYED, { _ }, SAMPLED, FORMAT, COMPONENTS>

pub fn gather<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, ARRAYED>, component: u32 ) -> SampledType::Vec4where Self: HasGather, F: Float,

Gathers the requested component from four texels.

pub fn sample<F>( &self, sampler: Sampler, coord: impl ImageCoordinate<F, DIM, ARRAYED> ) -> SampledType::SampleResultwhere F: Float,

Sample texels at coord from the image using sampler.

pub fn sample_bias<F>( &self, sampler: Sampler, coord: impl ImageCoordinate<F, DIM, ARRAYED>, bias: f32 ) -> SampledType::SampleResultwhere F: Float,

Sample texels at coord from the image using sampler, after adding the input bias to the implicit level of detail.

pub fn sample_by_lod<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, ARRAYED>, lod: f32 ) -> SampledType::SampleResultwhere F: Float,

Fetch a single texel with a sampler set at compile time Sample the image at a coordinate by a lod

pub fn sample_by_gradient<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, ARRAYED>, gradient_dx: impl ImageCoordinate<F, DIM, { _ }>, gradient_dy: impl ImageCoordinate<F, DIM, { _ }> ) -> SampledType::SampleResultwhere F: Float,

Sample the image based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

pub fn sample_depth_reference<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, ARRAYED>, depth_reference: f32 ) -> SampledTypewhere F: Float,

Sample the image’s depth reference

pub fn sample_depth_reference_by_lod<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, ARRAYED>, depth_reference: f32, lod: f32 ) -> SampledTypewhere F: Float,

Sample the image’s depth reference based on an explicit lod

pub fn sample_depth_reference_by_gradient<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, ARRAYED>, depth_reference: f32, gradient_dx: impl ImageCoordinate<F, DIM, { _ }>, gradient_dy: impl ImageCoordinate<F, DIM, { _ }> ) -> SampledTypewhere F: Float,

Sample the image’s depth reference based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const SAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, { _ }, { _ }, SAMPLED, FORMAT, COMPONENTS>

pub fn sample_with_project_coordinate<F>( &self, sampler: Sampler, project_coordinate: impl ImageCoordinate<F, DIM, { _ }> ) -> SampledType::SampleResultwhere F: Float,

Sample the image with a project coordinate

pub fn sample_with_project_coordinate_by_lod<F>( &self, sampler: Sampler, project_coordinate: impl ImageCoordinate<F, DIM, { _ }>, lod: f32 ) -> SampledType::Vec4where F: Float,

Sample the image with a project coordinate by a lod

pub fn sample_with_project_coordinate_by_gradient<F>( &self, sampler: Sampler, project_coordinate: impl ImageCoordinate<F, DIM, { _ }>, gradient_dx: impl ImageCoordinate<F, DIM, { _ }>, gradient_dy: impl ImageCoordinate<F, DIM, { _ }> ) -> SampledType::Vec4where F: Float,

Sample the image with a project coordinate based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

pub fn sample_depth_reference_with_project_coordinate<F>( &self, sampler: Sampler, project_coordinate: impl ImageCoordinate<F, DIM, { _ }>, depth_reference: f32 ) -> SampledTypewhere F: Float,

Sample the image’s depth reference with the project coordinate

pub fn sample_depth_reference_with_project_coordinate_by_lod<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, { _ }>, depth_reference: f32, lod: f32 ) -> SampledTypewhere F: Float,

Sample the image’s depth reference with the project coordinate based on an explicit lod

pub fn sample_depth_reference_with_project_coordinate_by_gradient<F>( &self, sampler: Sampler, coordinate: impl ImageCoordinate<F, DIM, { _ }>, depth_reference: f32, gradient_dx: impl ImageCoordinate<F, DIM, { _ }>, gradient_dy: impl ImageCoordinate<F, DIM, { _ }> ) -> SampledTypewhere F: Float,

Sample the image’s depth reference with the project coordinate based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const ARRAYED: u32, const MULTISAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { _ }, FORMAT, COMPONENTS>

pub fn read<I>( &self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED> ) -> SampledType::SampleResultwhere I: Integer,

Read a texel from an image without a sampler.

pub unsafe fn write<I, const N: usize>( &self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED>, texels: impl Vector<SampledType, N> )where I: Integer,

Write a texel to an image without a sampler.

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const FORMAT: u32, const ARRAYED: u32, const MULTISAMPLED: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { _ }, FORMAT, COMPONENTS>

pub fn read<I>( &self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED> ) -> SampledType::SampleResultwhere I: Integer,

Read a texel from an image without a sampler.

pub unsafe fn write<I, const N: usize>( &self, coordinate: impl ImageCoordinate<I, DIM, ARRAYED>, texels: impl Vector<SampledType, N> )where I: Integer,

Write a texel to an image without a sampler.

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DEPTH: u32, const ARRAYED: u32, const MULTISAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, { _ }, DEPTH, ARRAYED, MULTISAMPLED, { _ }, FORMAT, COMPONENTS>

pub fn read_subpass<I>( &self, coordinate: impl ImageCoordinateSubpassData<I, ARRAYED> ) -> SampledType::SampleResultwhere I: Integer,

Read a texel from subpass input attachment. Note: Vulkan only allows the read if the first two components of the coordinate are zero.

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const ARRAYED: u32, const MULTISAMPLED: u32, const SAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, SAMPLED, FORMAT, COMPONENTS>

pub fn query_levels(&self) -> u32where Self: HasQueryLevels,

Query the number of mipmap levels.

pub fn query_lod( &self, sampler: Sampler, coord: impl ImageCoordinate<f32, DIM, { _ }> ) -> SampledType::Vec2where Self: HasQueryLevels,

Query the mipmap level and the level of detail for a hypothetical sampling of Image at Coordinate using an implicit level of detail. The first component of the result contains the mipmap array layer. The second component of the result contains the implicit level of detail relative to the base level.

pub fn query_size<Size: ImageCoordinate<u32, DIM, ARRAYED> + Default>( &self ) -> Sizewhere Self: HasQuerySize,

Query the dimensions of Image, with no level of detail.

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DIM: u32, const DEPTH: u32, const ARRAYED: u32, const SAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, DIM, DEPTH, ARRAYED, { _ }, SAMPLED, FORMAT, COMPONENTS>

pub fn query_size_lod<Size: ImageCoordinate<u32, DIM, ARRAYED> + Default>( &self, lod: u32 ) -> Sizewhere Self: HasQuerySizeLod,

Query the dimensions of Image, with no level of detail.

impl<SampledType: SampleType<FORMAT, COMPONENTS>, const DEPTH: u32, const ARRAYED: u32, const SAMPLED: u32, const FORMAT: u32, const COMPONENTS: u32> Image<SampledType, { _ }, DEPTH, ARRAYED, { _ }, SAMPLED, FORMAT, COMPONENTS>

pub fn query_samples(&self) -> u32

Query the number of samples available per texel fetch in a multisample image.