spirv_std

Module indirect_command

Indirect command structs from vulkan

Structs

DispatchIndirectCommand

Structure specifying an indirect dispatching command

DrawIndexedIndirectCommand

Structure specifying an indexed indirect drawing command

DrawIndirectCommand

Structure specifying an indirect drawing command

DrawMeshTasksIndirectCommandEXT

Structure specifying a mesh tasks draw indirect command

TraceRaysIndirectCommand2KHR

Structure specifying the parameters of an indirect trace ray command with indirect shader binding tables

TraceRaysIndirectCommandKHR

Structure specifying the parameters of an indirect ray tracing command

Type Aliases

DeviceAddress

https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkDeviceAddress.html

DeviceSize

https://registry.khronos.org/vulkan/specs/1.3-extensions/man/html/VkDeviceSize.html