Module spirv_std::math

This math library is heavily borrowed from glam In the future we hope to be able to use it directly!

Re-exports

pub use mat2::*;

pub use mat3::*;

pub use mat4::*;

pub use vec2::*;

pub use vec3::*;

pub use vec4::*;

Modules

mat2

mat3

mat4

vec2

vec3

vec4

Traits

MathExt