Struct spirv_std::math::vec3::Vec3

#[repr(simd)]pub struct Vec3(pub f32, pub f32, pub f32);

A 3-dimensional vector without SIMD support.

Implementations

impl Vec3

pub const fn new(x: f32, y: f32, z: f32) -> Self

Creates a new Vec3.

pub const fn zero() -> Self

Creates a Vec3 with all elements set to 0.0.

pub const fn one() -> Self

Creates a Vec3 with all elements set to 1.0.

pub const fn unit_x() -> Self

Creates a Vec3 with values [x: 1.0, y: 0.0, z: 0.0].

pub const fn unit_y() -> Self

Creates a Vec3 with values [x: 0.0, y: 1.0, z: 0.0].

pub const fn unit_z() -> Self

Creates a Vec3 with values [x: 0.0, y: 0.0, z: 1.0].

pub const fn splat(v: f32) -> Self

Creates a Vec3 with all elements set to v.

pub fn extend(self, w: f32) -> Vec4

Creates a Vec4 from self and the given w value.

pub fn truncate(self) -> Vec2

Creates a Vec2 from the first three elements of self, removing z.

pub fn x(self) -> f32

Returns element x.

pub fn y(self) -> f32

Returns element y.

pub fn z(self) -> f32

Returns element z.

pub fn x_mut(&mut self) -> &mut f32

Returns a mutable reference to element x.

pub fn y_mut(&mut self) -> &mut f32

Returns a mutable reference to element y.

pub fn z_mut(&mut self) -> &mut f32

Returns a mutable reference to element z.

pub fn set_x(&mut self, x: f32)

Sets element x.

pub fn set_y(&mut self, y: f32)

Sets element y.

pub fn set_z(&mut self, z: f32)

Sets element z.

pub fn dot(self, other: Self) -> f32

Computes the dot product of self and other.

pub fn cross(self, other: Self) -> Self

Computes the cross product of self and other.

pub fn length(self) -> f32

Computes the length of self.

pub fn length_squared(self) -> f32

Computes the squared length of self.

This is generally faster than Vec3::length() as it avoids a square root operation.

pub fn length_reciprocal(self) -> f32

👎 Deprecated since 0.9.5:
please use Vec3::length_recip instead

pub fn length_recip(self) -> f32

Computes 1.0 / Vec3::length().

For valid results, self must not be of length zero.

pub fn normalize(self) -> Self

Returns self normalized to length 1.0.

For valid results, self must not be of length zero.

pub fn min(self, other: Self) -> Self

Returns the vertical minimum of self and other.

In other words, this computes [x: min(x1, x2), y: min(y1, y2), z: min(z1, z2)], taking the minimum of each element individually.

pub fn max(self, other: Self) -> Self

Returns the vertical maximum of self and other.

In other words, this computes [x: max(x1, x2), y: max(y1, y2), z: max(z1, z2)], taking the maximum of each element individually.

pub fn min_element(self) -> f32

Returns the horizontal minimum of self's elements.

In other words, this computes min(x, y, z).

pub fn max_element(self) -> f32

Returns the horizontal maximum of self's elements.

In other words, this computes max(x, y, z).

pub fn from_slice_unaligned(slice: &[f32]) -> Self

Creates a Vec3 from the first three values in slice.

Panics

Panics if slice is less than three elements long.

pub fn write_to_slice_unaligned(self, slice: &mut [f32])

Writes the elements of self to the first three elements in slice.

Panics

Panics if slice is less than three elements long.

pub fn abs(self) -> Self

Returns a Vec3 containing the absolute value of each element of self.

pub fn round(self) -> Self

Returns a Vec3 containing the nearest integer to a number for each element of self. Round half-way cases away from 0.0.

pub fn floor(self) -> Self

Returns a Vec3 containing the largest integer less than or equal to a number for each element of self.

pub fn pow(self, power: f32) -> Self

Returns a Vec3 containing this vector raised to the power of power

pub fn exp(self) -> Self

Returns a Vec3 containing this vector exp'd

pub fn ceil(self) -> Self

Returns a Vec3 containing the smallest integer greater than or equal to a number for each element of self.

pub fn reciprocal(self) -> Self

👎 Deprecated since 0.9.5:
please use Vec3::recip instead

pub fn recip(self) -> Self

Returns a Vec3 containing the reciprocal 1.0/n of each element of self.

pub fn lerp(self, other: Self, s: f32) -> Self

Performs a linear interpolation between self and other based on the value s.

When s is 0.0, the result will be equal to self. When s is 1.0, the result will be equal to other.