spirv_std::matrix

Struct Matrix4x3

#[repr(C)]
pub struct Matrix4x3 {
    pub x: Vec3A,
    pub y: Vec3A,
    pub z: Vec3A,
    pub w: Vec3A,
}

A Matrix with 4 columns of Vec3, very similar to glamโs Affine3A.

Primarily used in ray tracing extensions to represent object rotation, scale and translation.

Fields
x: Vec3Ay: Vec3Az: Vec3Aw: Vec3A
Implementations

impl Matrix4x3

The from_* fn signatures should match Affine3A, to make it easier to switch to Affine3A later. The to_* fn signatures are custom

pub fn from_affine3a(affine: Affine3A) -> Self

Convert from glamโs Affine3A

pub fn from_mat3(mat3: Mat3) -> Self

Creates an affine transform from a 3x3 matrix (expressing scale, shear and rotation)

pub fn from_mat3_translation(mat3: Mat3, translation: Vec3) -> Self

Creates an affine transform from a 3x3 matrix (expressing scale, shear and rotation) and a translation vector.

Equivalent to Affine3A::from_translation(translation) * Affine3A::from_mat3(mat3)

pub fn from_mat4(m: Mat4) -> Self

The given Mat4 must be an affine transform, i.e. contain no perspective transform.

pub fn to_affine3a(self) -> Affine3A

Convert to glamโs Affine3A

pub fn to_mat3a(self) -> Mat3A

Creates a 3x3 matrix representing the rotation and scale, cutting off the translation

pub fn to_mat3(self) -> Mat3

Creates a 3x3 matrix representing the rotation and scale, cutting off the translation

pub fn to_mat4(self) -> Mat4

Creates a 4x4 matrix from this affine transform