Enum spirv_std::memory::Scope

pub enum Scope {
    CrossDevice,
    Device,
    Workgroup,
    Subgroup,
    Invocation,
    QueueFamily,
}

Specification for how large of a scope some instructions should operate on - used when calling functions that take a configurable scope.

Variants

CrossDevice

Crosses multiple devices.

Device

The current device.

Workgroup

The current workgroup.

Subgroup

The current subgroup.

Invocation

The current invocation.

QueueFamily

The current queue family.