Enum spirv_std::memory::Semantics

pub enum Semantics {
    None,
    Acquire,
    Release,
    AcquireRelease,
    SequentiallyConsistent,
    UniformMemory,
    SubgroupMemory,
    WorkgroupMemory,
    CrossWorkgroupMemory,
    AtomicCounterMemory,
    ImageMemory,
    OutputMemory,
    MakeAvailable,
    MakeVisible,
    Volatile,
}

Variants

None

No memory semantics.

Acquire

On an atomic instruction, orders memory operations provided in program order after this atomic instruction against this atomic instruction. On a barrier, orders memory operations provided in program order after this barrier against atomic instructions before this barrier.

Release

On an atomic instruction, orders memory operations provided in program order before this atomic instruction against this atomic instruction. On a barrier, orders memory operations provided in program order before this barrier against atomic instructions after this barrier.

AcquireRelease

Has the properties of both Self::Acquire and Self::Release semantics. It is used for read-modify-write operations.

SequentiallyConsistent

All observers see this memory access in the same order with respect to other sequentially-consistent memory accesses from this invocation. If the declared memory model is vulkan, SequentiallyConsistent must not be used.

UniformMemory

Apply the memory-ordering constraints to [crate::storage_class::StorageBuffer], [crate::storage_class::PhysicalStorageBuffer], or [crate::storage_class::Uniform] Storage Class memory.

SubgroupMemory

Apply the memory-ordering constraints to subgroup memory.

WorkgroupMemory

Apply the memory-ordering constraints to [crate::storage_class::Workgroup] Storage Class memory.

CrossWorkgroupMemory

Apply the memory-ordering constraints to [crate::storage_class::CrossWorkgroup] Storage Class memory.

AtomicCounterMemory

Apply the memory-ordering constraints to [crate::storage_class::AtomicCounter] Storage Class memory.

ImageMemory

Apply the memory-ordering constraints to image contents (types declared by OpTypeImage), or to accesses done through pointers to the [crate::storage_class::Image] Storage Class.

OutputMemory

Apply the memory-ordering constraints to the [crate::storage_class::Output] Storage Class memory.

MakeAvailable

Perform an availability operation on all references in the selected storage classes.

MakeVisible

Perform a visibility operation on all references in the selected storage classes.

Volatile

This access cannot be eliminated, duplicated, or combined with other accesses.