spirv_std

Module ray_tracing

Ray-tracing data types

Structs

AccelerationStructure

An acceleration structure type which is an opaque reference to an acceleration structure handle as defined in the client API specification.

RayFlags

Flags controlling the properties of an OpTraceRayKHR instruction. Despite being a mask and allowing multiple bits to be combined, it is invalid for more than one of these four bits to be set: OPAQUE, NO_OPAQUE, CULL_OPAQUE, CULL_NO_OPAQUE, only one of CULL_BACK_FACING_TRIANGLES and CULL_FRONT_FACING_TRIANGLES may be set.

RayQuery

A ray query type which is an opaque object representing a ray traversal.

Enums

CandidateIntersection

Describes the type of the intersection which is currently the candidate in a ray query, returned by RayQuery::get_candidate_intersection_type.

CommittedIntersection

Describes the type of the intersection currently committed in a ray query, returned by RayQuery::get_committed_intersection_type.