Module spirv_std::storage_class

Storage Classes

Class of storage for declared variables. These types act as pointers to values either contained in the GPU's memory. For example; Input<f32> points to a f32 that was provided as input from the pipeline, and Private<f32> points to a f32 in the GPU's global memory. Intermediate values do not form a storage class, and unless stated otherwise, storage class-based restrictions are not restrictions on intermediate objects and their types.

Structs

AtomicCounter

CallableDataKHR

CrossWorkgroup

Function

Generic

HitAttributeKHR

Image

IncomingCallableDataKHR

IncomingRayPayloadKHR

Input

Output

PhysicalStorageBuffer

Private

PushConstant

RayPayloadKHR

ShaderRecordBufferKHR

StorageBuffer

Uniform

UniformConstant

Workgroup