Struct spirv_std::Cubemap

pub struct Cubemap { /* fields omitted */ }

👎 Deprecated:
Legacy image type. Use spirv_std::image::Cubemap instead.

Implementations

impl Cubemap

pub fn sample<V: Vector<f32, 4>>(
&self,
sampler: Sampler,
coordinate: impl Vector<f32, 3>
) -> V

pub fn sample_by_lod<V: Vector<f32, 4>>(
&self,
sampler: Sampler,
coordinate: impl Vector<f32, 3>,
lod: f32
) -> V

Sample the image at a coordinate by a lod

pub fn sample_by_gradient<V: Vector<f32, 4>>(
&self,
sampler: Sampler,
coordinate: impl Vector<f32, 3>,
gradient_dx: impl Vector<f32, 3>,
gradient_dy: impl Vector<f32, 3>
) -> V

Sample the image based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx, dw/dx], [du/dy, dv/dy, dw/dy])

pub fn sample_depth_reference(
&self,
sampler: Sampler,
coordinate: impl Vector<f32, 3>,
depth_reference: f32
) -> f32

Sample the image’s depth reference

pub fn sample_depth_reference_by_lod(
&self,
sampler: Sampler,
coordinate: impl Vector<f32, 3>,
depth_reference: f32,
lod: f32
) -> f32

Sample the image’s depth reference based on an explicit lod

pub fn sample_depth_reference_by_gradient(
&self,
sampler: Sampler,
coordinate: impl Vector<f32, 3>,
depth_reference: f32,
gradient_dx: impl Vector<f32, 3>,
gradient_dy: impl Vector<f32, 3>
) -> f32

Sample the image’s depth reference based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx, dw/dx], [du/dy, dv/dy, dw/dy])