spirv_std

Trait ScalarComposite

pub trait ScalarComposite:
    Copy
    + Send
    + Sync
    + 'static {
    // Required method
    fn transform<F: ScalarOrVectorTransform>(self, f: &mut F) -> Self;
}

A ScalarComposite is a type that is either

a Scalar

a Vector (since vectors are made from scalars)

an array of ScalarComposite

a struct where all members are ScalarComposite

an enum with a repr that is a Scalar

By calling Self::transform you can visit all the individual Scalar and Vector values this composite is build out of and transform them into some other value. This is particularly useful for subgroup intrinsics sending data to other threads.

To derive ScalarComposite on a struct, all members must also implement ScalarComposite.

To derive ScalarComposite on an enum, the enum must implement From<N> and Into<N> where N is defined by the #[repr(N)] attribute on the enum and must be an Integer, like u32. Note that some safe subgroup operations may return an “undefined result”, so your From<N> must gracefully handle arbitrary bit patterns being passed to it. While panicking is legal, it is discouraged as it may result in unexpected control flow. To implement these conversion traits, we recommend FromPrimitive and IntoPrimitive from the num_enum crate. FromPrimitive requires the enum to either be exhaustive or have a variant to default to, by either implementing Default or marking a variant with #[num_enum(default)]. Note to disable default features on the num_enum crate, or it won’t compile on SPIR-V.

Required Methods

fn transform<F: ScalarOrVectorTransform>(self, f: &mut F) -> Self

Transform the individual Scalar and Vector values of this type to a different value.

See Self for more detail.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.