spirv_std::vector

Trait VectorOrScalar

pub unsafe trait VectorOrScalar:
    Copy
    + Default
    + Send
    + Sync
    + 'static {
    type Scalar: Scalar;

    const DIM: NonZeroUsize;
}

Abstract trait representing either a vector or a scalar type.

Safety

Implementing this trait on non-scalar or non-vector types may break assumptions about other unsafe code, and should not be done.

Required Associated Constants

const DIM: NonZeroUsize

The dimension of the vector, or 1 if it is a scalar

Required Associated Types

type Scalar: Scalar

Either the scalar component type of the vector or the scalar itself.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.