Crate spirv_std

Re-exports

pub extern crate spirv_std_macros as macros;

pub use byte_addressable_buffer::ByteAddressableBuffer;

pub use num_traits;

pub use glam;

Modules

arch

SPIR-V Instrinics

byte_addressable_buffer

float

image

Image types

integer

memory

Types for handling memory ordering constraints for concurrent memory access.

ray_tracing

Ray-tracing data types

scalar

vector

Macros

Image

A macro for creating SPIR-V OpTypeImage types. Always produces a spirv_std::image::Image<...> type.

ray_query

Constructs an uninitialized ray query variable. Using the syntax let (mut)? <name>. Where name is the name of the ray query variable.

Structs

RuntimeArray

Sampler

An opaque reference to settings that describe how to access, filter, or sample an image.