Module spirv_std::arch

SPIR-V Intrinsics

This module is intended as a low level abstraction over SPIR-V instructions. These functions will typically map to a single instruction, and will perform no additional safety checks beyond type-checking.

Structs

SubgroupMask

SubgroupMask is a glam::UVec4 representing a bitmask of all invocations within a subgroup. Mostly used in group ballot operations.

Enums

GroupOperation

Defines the class of group operation.

QuadDirection

Direction is the kind of swap to perform.

Traits

IndexUnchecked

Index into an array without bounds checking.

Functions

all

Result is true if all components of vector is true, otherwise result is false.

all_memory_barrier⚠

Blocks execution of all threads in a group until all memory accesses have been completed.

all_memory_barrier_with_group_sync⚠

Blocks execution of all threads in a group until all memory accesses have been completed and all threads in the group have reached this call.

any

Result is true if any component of vector is true, otherwise result is false.

atomic_and⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_compare_exchange⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_exchange⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_f_add⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_f_max⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_f_min⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_i_add⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_i_decrement⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_i_increment⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_i_sub⚠

Perform the following steps atomically with respect to any other atomic accesses within SCOPE to the same location:

atomic_load⚠

Atomically load through ptr using the given SEMANTICS. All subparts of the value that is loaded are read atomically with respect to all other atomic accesses to it within SCOPE.

atomic_or⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_s_max⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_s_min⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_store⚠

Atomically store through ptr using the given SEMANTICS. All subparts of value are written atomically with respect to all other atomic accesses to it within SCOPE.

atomic_u_max⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_u_min⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

atomic_xor⚠

Perform the following steps atomically with respect to any other atomic accesses within Scope to the same location:

control_barrier⚠

Wait for other invocations of this module to reach the current point of execution.

ddx

Returns the partial derivative of component with respect to the window’s X coordinate. Returns the same result as either ddx_fine or ddx_coarse, selection of which one is dependent on external factors.

ddx_coarse

Returns the partial derivative of component with respect to the window’s X coordinate. Uses local differencing based on the value of component for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of component for the current fragment. That is, over a given area, the implementation can compute X derivatives in fewer unique locations than would be allowed by ddx_fine.

ddx_coarse_vector

Returns the partial derivative of component with respect to the window’s X coordinate. Uses local differencing based on the value of component for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of component for the current fragment. That is, over a given area, the implementation can compute X derivatives in fewer unique locations than would be allowed by ddx_fine.

ddx_fine

Returns the partial derivative of component with respect to the window’s X coordinate. Uses local differencing based on the value of component for the current fragment and its immediate neighbor(s).

ddx_fine_vector

Returns the partial derivative of component with respect to the window’s X coordinate. Uses local differencing based on the value of component for the current fragment and its immediate neighbor(s).

ddx_vector

Returns the partial derivative of component with respect to the window’s X coordinate. Returns the same result as either ddx_fine or ddx_coarse, selection of which one is dependent on external factors.

ddy

Returns the partial derivative of component with respect to the window’s Y coordinate. Returns the same result as either ddy_fine or ddy_coarse, selection of which one is dependent on external factors.

ddy_coarse

Returns the partial derivative of component with respect to the window’s Y coordinate. Uses local differencing based on the value of component for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of component for the current fragment. That is, over a given area, the implementation can compute Y derivatives in fewer unique locations than would be allowed by ddy_fine.

ddy_coarse_vector

Returns the partial derivative of component with respect to the window’s Y coordinate. Uses local differencing based on the value of component for the current fragment’s neighbors, and possibly, but not necessarily, includes the value of component for the current fragment. That is, over a given area, the implementation can compute Y derivatives in fewer unique locations than would be allowed by ddy_fine.

ddy_fine

Returns the partial derivative of component with respect to the window’s Y coordinate. Uses local differencing based on the value of component for the current fragment and its immediate neighbor(s).

ddy_fine_vector

Returns the partial derivative of component with respect to the window’s Y coordinate. Uses local differencing based on the value of component for the current fragment and its immediate neighbor(s).

ddy_vector

Returns the partial derivative of component with respect to the window’s Y coordinate. Returns the same result as either ddy_fine or ddy_coarse, selection of which one is dependent on external factors.

demote_to_helper_invocation⚠

Demote fragment shader invocation to a helper invocation. Equivalvent to discard() in HLSL. Any stores to memory after this instruction are suppressed and the fragment does not write outputs to the framebuffer.

device_memory_barrier⚠

Blocks execution of all threads in a group until all device memory accesses have been completed.

device_memory_barrier_with_group_sync⚠

Blocks execution of all threads in a group until all device memory accesses have been completed and all threads in the group have reached this call.

emit_stream_vertex⚠

Emits the current values of all output variables to the current output primitive. After execution, the values of all output variables are undefined.

emit_vertex⚠

Emits the current values of all output variables to the current output primitive. After execution, the values of all output variables are undefined. Requires capability Geometry.

end_primitive⚠

Finish the current primitive and start a new one. No vertex is emitted. Requires capability Geometry.

end_stream_primitive⚠

Finish the current primitive and start a new one. No vertex is emitted.

execute_callable⚠

Invoke a callable shader.

fwidth

Returns the sum of the absolute values of ddx and ddy as a single operation.

fwidth_coarse

Returns the sum of the absolute values of ddx_coarse and ddy_coarse as a single operation.

fwidth_coarse_vector

Returns the sum of the absolute values of ddx_coarse and ddy_coarse as a single operation.

fwidth_fine

Returns the sum of the absolute values of ddx_fine and ddy_fine as a single operation.

fwidth_fine_vector

Returns the sum of the absolute values of ddx_fine and ddy_fine as a single operation.

fwidth_vector

Returns the sum of the absolute values of ddx and ddy as a single operation.

ignore_intersection⚠

Ignores the current potential intersection, terminating the invocation that executes it, and continues the ray traversal. This instruction is allowed only in any_hit execution model.

is_helper_invocation

Returns true if the invocation is currently a helper invocation, otherwise result is false. An invocation is currently a helper invocation if it was originally invoked as a helper invocation or if it has been demoted to a helper invocation by demote_to_helper_invocation.

kill

Fragment-shader discard. Equivalvent to discard() from GLSL

memory_barrier⚠

Control the order that memory accesses are observed.

report_intersection⚠

Reports an intersection back to the traversal infrastructure.

signed_max

Compute the maximum of two signed integers via a GLSL extended instruction.

signed_min

Compute the minimum of two signed integers via a GLSL extended instruction.

subgroup_all⚠

Evaluates a predicate for all active invocations in the group, resulting in true if predicate evaluates to true for all active invocations in the group, otherwise the result is false.

subgroup_all_equal⚠

Evaluates a value for all active invocations in the group. The result is true if Value is equal for all active invocations in the group. Otherwise, the result is false.

subgroup_and⚠

A bitwise and group operation of all Value operands contributed by active invocations in the group.

subgroup_any⚠

Evaluates a predicate for all active invocations in the group, resulting in true if predicate evaluates to true for any active invocation in the group, otherwise the result is false.

subgroup_ballot⚠

Result is a bitfield value combining the Predicate value from all invocations in the group that execute the same dynamic instance of this instruction. The bit is set to one if the corresponding invocation is active and the Predicate for that invocation evaluated to true; otherwise, it is set to zero.

subgroup_ballot_bit_count⚠

Result is the number of bits that are set to 1 in Value, considering only the bits in Value required to represent all bits of the group’s invocations.

subgroup_ballot_bit_extract⚠

Evaluates a value for all active invocations in the group, resulting in true if the bit in Value that corresponds to Index is set to one, otherwise the result is false.

subgroup_ballot_exclusive_bit_count⚠

Result is the number of bits that are set to 1 in Value, considering only the bits in Value required to represent all bits of the group’s invocations.

subgroup_ballot_find_lsb⚠

Find the least significant bit set to 1 in Value, considering only the bits in Value required to represent all bits of the group’s invocations. If none of the considered bits is set to 1, the resulting value is undefined.

subgroup_ballot_find_msb⚠

Find the most significant bit set to 1 in Value, considering only the bits in Value required to represent all bits of the group’s invocations. If none of the considered bits is set to 1, the resulting value is undefined.

subgroup_ballot_inclusive_bit_count⚠

Result is the number of bits that are set to 1 in Value, considering only the bits in Value required to represent all bits of the group’s invocations.

subgroup_barrier⚠

Only usable if the extension GL_KHR_shader_subgroup_basic is enabled.

subgroup_broadcast⚠

Result is the Value of the invocation identified by the id Id to all active invocations in the group.

subgroup_broadcast_first⚠

Result is the Value of the invocation from the active invocation with the lowest id in the group to all active invocations in the group.

subgroup_clustered_and⚠

A bitwise and group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_f_add⚠

A floating point add group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_f_max⚠

A floating point maximum group operation of all Value operands contributed by active invocations in by group.

subgroup_clustered_f_min⚠

A floating point minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_f_mul⚠

A floating point multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_i_add⚠

An integer add group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_i_mul⚠

An integer multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_logical_and⚠

A logical and group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_logical_or⚠

A logical or group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_logical_xor⚠

A logical xor group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_or⚠

A bitwise or group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_s_max⚠

A signed integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_s_min⚠

A signed integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_u_max⚠

An unsigned integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_u_min⚠

An unsigned integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_clustered_xor⚠

A bitwise xor group operation of all Value operands contributed by active invocations in the group.

subgroup_elect⚠

Result is true only in the active invocation with the lowest id in the group, otherwise result is false.

subgroup_exclusive_and⚠

A bitwise and group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_f_add⚠

A floating point add group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_f_max⚠

A floating point maximum group operation of all Value operands contributed by active invocations in by group.

subgroup_exclusive_f_min⚠

A floating point minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_f_mul⚠

A floating point multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_i_add⚠

An integer add group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_i_mul⚠

An integer multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_logical_and⚠

A logical and group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_logical_or⚠

A logical or group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_logical_xor⚠

A logical xor group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_or⚠

A bitwise or group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_s_max⚠

A signed integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_s_min⚠

A signed integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_u_max⚠

An unsigned integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_u_min⚠

An unsigned integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_exclusive_xor⚠

A bitwise xor group operation of all Value operands contributed by active invocations in the group.

subgroup_f_add⚠

A floating point add group operation of all Value operands contributed by active invocations in the group.

subgroup_f_max⚠

A floating point maximum group operation of all Value operands contributed by active invocations in by group.

subgroup_f_min⚠

A floating point minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_f_mul⚠

A floating point multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_i_add⚠

An integer add group operation of all Value operands contributed by active invocations in the group.

subgroup_i_mul⚠

An integer multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_and⚠

A bitwise and group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_f_add⚠

A floating point add group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_f_max⚠

A floating point maximum group operation of all Value operands contributed by active invocations in by group.

subgroup_inclusive_f_min⚠

A floating point minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_f_mul⚠

A floating point multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_i_add⚠

An integer add group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_i_mul⚠

An integer multiply group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_logical_and⚠

A logical and group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_logical_or⚠

A logical or group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_logical_xor⚠

A logical xor group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_or⚠

A bitwise or group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_s_max⚠

A signed integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_s_min⚠

A signed integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_u_max⚠

An unsigned integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_u_min⚠

An unsigned integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_inclusive_xor⚠

A bitwise xor group operation of all Value operands contributed by active invocations in the group.

subgroup_inverse_ballot⚠

Evaluates a value for all active invocations in the group, resulting in true if the bit in Value for the corresponding invocation is set to one, otherwise the result is false.

subgroup_logical_and⚠

A logical and group operation of all Value operands contributed by active invocations in the group.

subgroup_logical_or⚠

A logical or group operation of all Value operands contributed by active invocations in the group.

subgroup_logical_xor⚠

A logical xor group operation of all Value operands contributed by active invocations in the group.

subgroup_memory_barrier⚠

Only usable if the extension GL_KHR_shader_subgroup_basic is enabled.

subgroup_memory_barrier_buffer⚠

Only usable if the extension GL_KHR_shader_subgroup_basic is enabled.

subgroup_memory_barrier_image⚠

Only usable if the extension GL_KHR_shader_subgroup_basic is enabled.

subgroup_memory_barrier_shared⚠

Only usable if the extension GL_KHR_shader_subgroup_basic is enabled.

subgroup_or⚠

A bitwise or group operation of all Value operands contributed by active invocations in the group.

subgroup_quad_broadcast⚠

Result is the Value of the invocation within the quad with a quad index equal to Index.

subgroup_quad_swap⚠

Swap the Value of the invocation within the quad with another invocation in the quad using Direction.

subgroup_s_max⚠

A signed integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_s_min⚠

A signed integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_shuffle⚠

Result is the Value of the invocation identified by the id Id.

subgroup_shuffle_down⚠

Result is the Value of the invocation identified by the current invocation’s id within the group + Delta.

subgroup_shuffle_up⚠

Result is the Value of the invocation identified by the current invocation’s id within the group - Delta.

subgroup_shuffle_xor⚠

Result is the Value of the invocation identified by the current invocation’s id within the group xor’ed with Mask.

subgroup_u_max⚠

An unsigned integer maximum group operation of all Value operands contributed by active invocations in the group.

subgroup_u_min⚠

An unsigned integer minimum group operation of all Value operands contributed by active invocations in the group.

subgroup_xor⚠

A bitwise xor group operation of all Value operands contributed by active invocations in the group.

terminate_ray⚠

Terminates the invocation that executes it, stops the ray traversal, accepts the current hit, and invokes the closest_hit execution model (if active). This instruction is allowed only in the any_hit execution model.

unsigned_max

Compute the maximum of two unsigned integers via a GLSL extended instruction.

unsigned_min

Compute the minimum of two unsigned integers via a GLSL extended instruction.

vector_extract_dynamic⚠

Extract a single, dynamically selected, component of a vector.

vector_insert_dynamic⚠

Make a copy of a vector, with a single, variably selected, component modified.

workgroup_memory_barrier⚠

Blocks execution of all threads in a group until all group shared accesses have been completed.

workgroup_memory_barrier_with_group_sync⚠

Blocks execution of all threads in a group until all group shared accesses have been completed and all threads in the group have reached this call.