spirv_std::arch

Function set_mesh_outputs_ext

pub unsafe fn set_mesh_outputs_ext(vertex_count: u32, primitive_count: u32)

Sets the actual output size of the primitives and vertices that the mesh shader workgroup will emit upon completion.

‘Vertex Count’ must be a 32-bit unsigned integer value. It defines the array size of per-vertex outputs.

‘Primitive Count’ must a 32-bit unsigned integer value. It defines the array size of per-primitive outputs.

The arguments are taken from the first invocation in each workgroup. Any invocation must execute this instruction no more than once and under uniform control flow. There must not be any control flow path to an output write that is not preceded by this instruction.

This instruction is only valid in the MeshEXT Execution Model.