spirv_std::byte_addressable_buffer

Struct ByteAddressableBuffer

#[repr(transparent)]
pub struct ByteAddressableBuffer<T> {
    pub data: T,
}

ByteAddressableBuffer is a view to an untyped blob of data, allowing loads and stores of arbitrary basic data types at arbitrary indices.

Alignment

All data must be aligned to size 4, each element within the data (e.g. struct fields) must have a size and alignment of a multiple of 4, and the byte_index passed to load and store must be a multiple of 4. Technically it is not a byte addressable buffer, but rather a word buffer, but this naming and behavior was inherited from HLSL (where it’s UB to pass in an index not a multiple of 4).

Safety

Using these functions allows reading a different type from the buffer than was originally written (by a previous store() or the host API), allowing all sorts of safety guarantees to be bypassed, making it effectively a transmute.

Fields
data: T

The underlying array of bytes, able to be directly accessed.

Implementations

impl<'a> ByteAddressableBuffer<&'a [u32]>

pub fn from_slice(data: &'a [u32]) -> Self

Creates a ByteAddressableBuffer from the untyped blob of data.

pub unsafe fn load<T>(&self, byte_index: u32) -> T

Loads an arbitrary type from the buffer. byte_index must be a multiple of 4.

Safety

See Self.

pub unsafe fn load_unchecked<T>(&self, byte_index: u32) -> T

Loads an arbitrary type from the buffer. byte_index must be a multiple of 4.

Safety

See Self. Additionally, bounds or alignment checking is not performed.

impl<'a> ByteAddressableBuffer<&'a mut [u32]>

pub fn from_mut_slice(data: &'a mut [u32]) -> Self

Creates a ByteAddressableBuffer from the untyped blob of data.

pub fn as_ref(&self) -> ByteAddressableBuffer<&[u32]>

Create a non-mutable ByteAddressableBuffer from this mutable one.

pub unsafe fn load<T>(&self, byte_index: u32) -> T

Loads an arbitrary type from the buffer. byte_index must be a multiple of 4.

Safety

See Self.

pub unsafe fn load_unchecked<T>(&self, byte_index: u32) -> T

Loads an arbitrary type from the buffer. byte_index must be a multiple of 4.

Safety

See Self. Additionally, bounds or alignment checking is not performed.

pub unsafe fn store<T>(&mut self, byte_index: u32, value: T)

Stores an arbitrary type into the buffer. byte_index must be a multiple of 4.

Safety

See Self.

pub unsafe fn store_unchecked<T>(&mut self, byte_index: u32, value: T)

Stores an arbitrary type into the buffer. byte_index must be a multiple of 4.

Safety

See Self. Additionally, bounds or alignment checking is not performed.