Module spirv_std::image

Image types

Macros

Image
A macro for creating SPIR-V OpTypeImage types.

Structs

Image
An opaque image type. Corresponds to OpTypeImage.

SampledImage
An image combined with a sampler, enabling filtered accesses of the image’s contents.

Enums

AccessQualifier
The access permissions for the image.

Arrayed
Whether the image uses arrayed content.

Dimensionality
The dimension of the image.

ImageDepth
Whether a given image contains depth information. Note Whether or not to perform depth comparisons is a property of the sampling code, not of this type.

ImageFormat
The underlying internal representation of the image.

Multisampled
Whether the image uses arrayed content.

Sampled
Whether or not the image will be accessed in combination with a sampler.

Traits

ImageCoordinate
Marker trait for arguments that accept a coordinate for an crate::Image.

ImageCoordinateSubpassData
Marker trait for arguments that are valid for a crate::image::Dimensionality::SubpassData image query.

SampleType
Marker trait for arguments that accept single scalar values or vectors of scalars.

Type Definitions

Cubemap

Image2d

Image2dArray

StorageImage2d