Enum spirv_std::image::ImageFormat

#[repr(u32)]
pub enum ImageFormat {

    Unknown,
    Rgba32f,
    Rgba16f,
    R32f,
    Rgba8,
    Rgba8Snorm,
    Rg32f,
    Rg16f,
    R11fG11fB10f,
    R16f,
    Rgba16,
    Rgb10A2,
    Rg16,
    Rg8,
    R16,
    R8,
    Rgba16Snorm,
    Rg16Snorm,
    Rg8Snorm,
    R16Snorm,
    R8Snorm,
    Rgba32i,
    Rgba16i,
    Rgba8i,
    R32i,
    Rg32i,
    Rg16i,
    Rg8i,
    R16i,
    R8i,
    Rgba32ui,
    Rgba16ui,
    Rgba8ui,
    R32ui,
    Rgb10A2ui,
    Rg32ui,
    Rg16ui,
    Rg8ui,
    R16ui,
    R8ui,
    R64ui,
    R64i,

}

The underlying internal representation of the image.

Variants

Unknown

Representation not known at compile time.

Rgba32f

RGBA channels, 32 bit floating point integer.

Rgba16f

RGBA channels, 16 bit floating point integer.

R32f

RGBA channels, 16 bit floating point integer.

Rgba8

RGBA channels, 8 bit floating point integer.

Rgba8Snorm

RGBA channels, 8 bit signed normalized integer.

Rg32f

Red+Green channels, 32 bit floating point integer.

Rg16f

Red+Green channels, 16 bit floating point integer.

R11fG11fB10f

32 bit unsigned integer containing two 11 bit floating point integers for the Red and Green channels, and a 10 bit floating point integer for the Blue channel.

R16f

Red channel, 16 bit floating point.

Rgba16

RGBA channel, 16 bit floating point.

Rgb10A2

32 bit unsigned integer containing three 10 bit unsigned normalized integers for the Red, Green, and Blue channels; with a 2 unsigned normalized integer for the Alpha channel.

Rg16

Red+Green channels, 16 bit floating point integer.

Rg8

Red+Green channels, 8 bit floating point integer.

R16

Red+Green channels, 16 bit floating point integer.

R8

Red channel, 8 bit floating point integer.

Rgba16Snorm

RGBA channels, 16 bit signed normalized integer.

Rg16Snorm

RGB channels, 16 bit signed normalized integer.

Rg8Snorm

Red+Green channels, 8 bit signed normalized integer.

R16Snorm

Red channel, 16 bit signed normalized integer.

R8Snorm

Red channel, 16 bit signed normalized integer.

Rgba32i

RGBA channels, 32 bit signed integer.

Rgba16i

RGBA channels, 16 bit signed integer.

Rgba8i

RGBA channels, 8 bit signed integer.

R32i

Red channel, 32 bit signed integer.

Rg32i

Red+Green channels, 32 bit signed integer.

Rg16i

Red+Green channels, 16 bit signed integer.

Rg8i

Red+Green channels, 8 bit signed integer.

R16i

Red channel, 16 bit signed integer.

R8i

Red channel, 8 bit signed integer.

Rgba32ui

RGBA channels, 32 bit unsigned integer.

Rgba16ui

RGBA channels, 16 bit unsigned integer.

Rgba8ui

RGBA channels, 8 bit unsigned integer.

R32ui

Red channel, 32 bit unsigned integer.

Rgb10A2ui

32 bit unsigned integer containing three 10 bit unsigned integers for the Red, Green, and Blue channels, and a 2 bit unsigned integer for the Alpha channel.

Rg32ui

Red+Green channels, 32 bit unsigned integer.

Rg16ui

Red+Green channels, 16 bit unsigned integer.

Rg8ui

Red+Green channels, 8 bit unsigned integer.

R16ui

Red channel, 16 bit unsigned integer.

R8ui

Red channel, 8 bit unsigned integer.

R64ui

Red channel, 64 bit unsigned integer.

R64i

Red channel, 64 bit signed integer.