Struct spirv_std::image::Image

pub struct Image<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const MULTISAMPLED: Multisampled, const SAMPLED: Sampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> { /* fields omitted */ }

An opaque image type. Corresponds to OpTypeImage.

You likely want to write this type using the crate::Image! macro helper, as the generic arguments here can get extremely verbose.

Implementations

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const MULTISAMPLED: Multisampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::Yes }, FORMAT, ACCESS_QUALIFIER>

pub fn fetch<V, I>(
&self,
coordinate: impl ImageCoordinate<I, DIM, ARRAYED>
) -> V where
V: Vector<SampledType, 4>,
I: Integer,

Fetch a single texel with a sampler set at compile time

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const FORMAT: ImageFormat, const ARRAYED: Arrayed, const SAMPLED: Sampled, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, ARRAYED, { Multisampled::False }, SAMPLED, FORMAT, ACCESS_QUALIFIER>

pub fn gather<F, V>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
component: u32
) -> V where
Self: HasGather,
F: Float,
V: Vector<SampledType, 4>,

Gathers the requested component from four texels.

pub fn sample<F, V>(
&self,
sampler: Sampler,
coord: impl ImageCoordinate<F, DIM, ARRAYED>
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Sample texels at coord from the image using sampler.

pub fn sample_bias<F, V>(
&self,
sampler: Sampler,
coord: impl ImageCoordinate<F, DIM, ARRAYED>,
bias: f32
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Sample texels at coord from the image using sampler, after adding the input bias to the implicit level of detail.

pub fn sample_by_lod<F, V>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
lod: f32
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Fetch a single texel with a sampler set at compile time Sample the image at a coordinate by a lod

pub fn sample_by_gradient<F, V>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
gradient_dx: impl ImageCoordinate<F, DIM, { Arrayed::False }>,
gradient_dy: impl ImageCoordinate<F, DIM, { Arrayed::False }>
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Sample the image based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

pub fn sample_depth_reference<F>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
depth_reference: f32
) -> SampledType where
F: Float,

Sample the image’s depth reference

pub fn sample_depth_reference_by_lod<F>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
depth_reference: f32,
lod: f32
) -> SampledType where
F: Float,

Sample the image’s depth reference based on an explicit lod

pub fn sample_depth_reference_by_gradient<F>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, ARRAYED>,
depth_reference: f32,
gradient_dx: impl ImageCoordinate<F, DIM, { Arrayed::False }>,
gradient_dy: impl ImageCoordinate<F, DIM, { Arrayed::False }>
) -> SampledType where
F: Float,

Sample the image’s depth reference based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const SAMPLED: Sampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, { Arrayed::False }, { Multisampled::False }, SAMPLED, FORMAT, ACCESS_QUALIFIER>

pub fn sample_with_project_coordinate<F, V>(
&self,
sampler: Sampler,
project_coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True }>
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Sample the image with a project coordinate

pub fn sample_with_project_coordinate_by_lod<F, V>(
&self,
sampler: Sampler,
project_coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True }>,
lod: f32
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Sample the image with a project coordinate by a lod

pub fn sample_with_project_coordinate_by_gradient<F, V>(
&self,
sampler: Sampler,
project_coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True }>,
gradient_dx: impl ImageCoordinate<F, DIM, { Arrayed::False }>,
gradient_dy: impl ImageCoordinate<F, DIM, { Arrayed::False }>
) -> V where
F: Float,
V: Vector<SampledType, 4>,

Sample the image with a project coordinate based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

pub fn sample_depth_reference_with_project_coordinate<F>(
&self,
sampler: Sampler,
project_coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True }>,
depth_reference: f32
) -> SampledType where
F: Float,

Sample the image’s depth reference with the project coordinate

pub fn sample_depth_reference_with_project_coordinate_by_lod<F>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True }>,
depth_reference: f32,
lod: f32
) -> SampledType where
F: Float,

Sample the image’s depth reference with the project coordinate based on an explicit lod

pub fn sample_depth_reference_with_project_coordinate_by_gradient<F>(
&self,
sampler: Sampler,
coordinate: impl ImageCoordinate<F, DIM, { Arrayed::True }>,
depth_reference: f32,
gradient_dx: impl ImageCoordinate<F, DIM, { Arrayed::False }>,
gradient_dy: impl ImageCoordinate<F, DIM, { Arrayed::False }>
) -> SampledType where
F: Float,

Sample the image’s depth reference with the project coordinate based on a gradient formed by (dx, dy). Specifically, ([du/dx, dv/dx], [du/dy, dv/dy])

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const MULTISAMPLED: Multisampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::No }, FORMAT, ACCESS_QUALIFIER>

pub fn read<I, V, const N: usize>(
&self,
coordinate: impl ImageCoordinate<I, DIM, ARRAYED>
) -> V where
I: Integer,
V: Vector<SampledType, N>,

Read a texel from an image without a sampler.

pub unsafe fn write<I, const N: usize>(
&self,
coordinate: impl ImageCoordinate<I, DIM, ARRAYED>,
texels: impl Vector<SampledType, N>
) where
I: Integer,

Write a texel to an image without a sampler.

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const FORMAT: ImageFormat, const ARRAYED: Arrayed, const MULTISAMPLED: Multisampled, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::Unknown }, FORMAT, ACCESS_QUALIFIER>

pub fn read<I, V, const N: usize>(
&self,
coordinate: impl ImageCoordinate<I, DIM, ARRAYED>
) -> V where
I: Integer,
V: Vector<SampledType, N>,

Read a texel from an image without a sampler.

pub unsafe fn write<I, const N: usize>(
&self,
coordinate: impl ImageCoordinate<I, DIM, ARRAYED>,
texels: impl Vector<SampledType, N>
) where
I: Integer,

Write a texel to an image without a sampler.

impl<SampledType: SampleType<FORMAT>, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const MULTISAMPLED: Multisampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, { Dimensionality::SubpassData }, DEPTH, ARRAYED, MULTISAMPLED, { Sampled::No }, FORMAT, ACCESS_QUALIFIER>

pub fn read_subpass<I, V, const N: usize>(
&self,
coordinate: impl ImageCoordinateSubpassData<I, ARRAYED>
) -> V where
I: Integer,
V: Vector<SampledType, N>,

Read a texel from subpass input attachment. Note: Vulkan only allows the read if the first two components of the coordinate are zero.

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const MULTISAMPLED: Multisampled, const SAMPLED: Sampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, ARRAYED, MULTISAMPLED, SAMPLED, FORMAT, ACCESS_QUALIFIER>

pub fn query_levels(&self) -> u32 where
Self: HasQueryLevels,

Query the number of mipmap levels.

pub fn query_lod<V: Vector<f32, 2>>(
&self,
sampler: Sampler,
coord: impl ImageCoordinate<f32, DIM, { Arrayed::False }>
) -> V where
Self: HasQueryLevels,

Query the mipmap level and the level of detail for a hypothetical sampling of Image at Coordinate using an implicit level of detail. The first component of the result contains the mipmap array layer. The second component of the result contains the implicit level of detail relative to the base level.

pub fn query_size<Size: ImageCoordinate<u32, DIM, ARRAYED> + Default>(
&self
) -> Size where
Self: HasQuerySize,

Query the dimensions of Image, with no level of detail.

impl<SampledType: SampleType<FORMAT>, const DIM: Dimensionality, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const SAMPLED: Sampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, DIM, DEPTH, ARRAYED, { Multisampled::False }, SAMPLED, FORMAT, ACCESS_QUALIFIER>

pub fn query_size_lod<Size: ImageCoordinate<u32, DIM, ARRAYED> + Default>(
&self,
lod: u32
) -> Size where
Self: HasQuerySizeLod,

Query the dimensions of Image, with no level of detail.

impl<SampledType: SampleType<FORMAT>, const DEPTH: ImageDepth, const ARRAYED: Arrayed, const SAMPLED: Sampled, const FORMAT: ImageFormat, const ACCESS_QUALIFIER: Option<AccessQualifier>> Image<SampledType, { Dimensionality::TwoD }, DEPTH, ARRAYED, { Multisampled::True }, SAMPLED, FORMAT, ACCESS_QUALIFIER>

pub fn query_samples(&self) -> u32

Query the number of samples available per texel fetch in a multisample image.