Trait spirv_std::image::HasGather

pub trait HasGather { }

This is a marker trait to represent the constraints on OpImageGather too complex to be represented by const generics. Specifically:

“Its OpTypeImage must have a Dim of 2D, Cube, or Rect. The MS operand of the underlying OpTypeImage must be 0.”