Trait spirv_std::image::HasQueryLevels

pub trait HasQueryLevels { }

This is a marker trait to represent the constraints on OpImageQueryLevels and OpImageQueryLod too complex to be represented by const generics. Specifically:

“Its Dim operand must be one of 1D, 2D, 3D, or Cube.”