Trait spirv_std::image::HasQuerySize

pub trait HasQuerySize { }

This is a marker trait to represent the constraints on OpImageQuerySize too complex to be represented by const generics. Specifically:

“Its Dim operand must be 1D, 2D, 3D, Cube, Rect, or Buffer. Additionally, if its Dim is 1D, 2D, 3D, or Cube, it must also have either an MS of 1 or a Sampled of 0 or 2.”