Trait spirv_std::image::HasQuerySizeLod

pub trait HasQuerySizeLod { }

This is a marker trait to represent the constraints on OpImageQuerySizeLod too complex to be represented by const generics. Specifically:

“Its Dim operand must be one of 1D, 2D, 3D, or Cube, and its MS must be 0.”