spirv_std::vector

Trait VectorTruncateInto

pub trait VectorTruncateInto<T> {
    // Required method
    fn truncate_into(self) -> T;
}

Trait that implements slicing of a vector into a scalar or vector of lower dimensions, by ignoring the higher dimensions

Required Methods

fn truncate_into(self) -> T

Slices the vector into a lower dimensional type by ignoring the higher components